More unit tests
