+++ lib/wlib/stl/Equal.h
